requirements.txt
requests
google-generativeai
certifi
pdf2image
Pillow